# swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "description": "This is a simple API that allows you to register products, categories and orders",
        "title": "Food Menu API",
        "contact": {
            "email": "rafael@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Test API"
        },
        {
            "url": "http://localhost:4000",
            "description": "Auth API"
        },       
        {
            "url": "http://foodapiexample.com",
            "description": "Production API"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "summary": "Generate Access Token",
                "description": "Generate Access token. Add the email and before clicking Execute, make sure to select the **AUTH API** option in the ***Servers*** menu",
                "tags": ["Token"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema":{
                                "$ref": "#/components/schemas/TokenRequest"
                            },
                            "examples": {
                                "token": {
                                    "email": "example@email.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Access Denied"
                    },
                    "404":{
                        "description": "Not Found. Check If you are using the correct server!"
                    },
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "accessToken": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/token": {
            "post": {
                "summary": "Refresh Token",
                "description": "Refresh Token. Add the refresh token in order to get a new Access Token and before clicking Execute, make sure to select the **AUTH API** option in the ***Servers*** menu",
                "tags": ["Token"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema":{
                                "$ref":"#/components/schemas/RefreshTokenRequest"
                            },
                            "examples": {
                                "token": {
                                    "email": "YOUR REFRESH TOKEN"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Access Denied"
                    },
                    "404":{
                        "description": "Not Found. Check If you are using the correct server!"
                    },
                    "200":{
                        "description": "OK",
                        "content": {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "accessToken": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TokenRequest": {
                "type": "object",
                "properties": {
                    "email": "string"
                }
            },
            "RefreshTokenRequest": {
                "type": "object",
                "properties": {
                    "token": "string"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}